+++ src/components/InfographicEditor/Toolbar/ToolbarContent.js
@@ -130,7 +130,7 @@ class ToolbarContent extends React.Component
             return this._CreateTextIcons();
         } else if (this.props.options === 'header-editor') {
             return this._CreateHeaderIcons();
-        } else if (this.props.options === 'image-editor') {
+        } else if (this.props.options === 'image-editor' || this.props.options === 'insert-image') {
             return this._CreateImageIcons();
         } else if (this.props.options === 'icon-editor') {
             return this._CreateIconIcons();
@@ -286,13 +286,21 @@ class ToolbarContent extends React.Component
 
     _CreateImageIcons()
     {
+        let editorClassName = 'insert-button' + (this.props.currentEditor === 
+            'image-editor' ? ' selected-editor' : ''),
+            replaceImageClassName = 'insert-button' + (this.props.currentEditor === 'update-image' ? ' selected-editor' : '');
         return (
             <div className='chart-container'>
                 {this._CreateCanvasSettings()}
                 <div className='editor-settings'>
                     <button 
                         id='editor-button' 
-                        onClick={() => { }}>
+                        className={editorClassName}
+                        onClick={() => { 
+                            if (this.props.currentEditor === 'update-image') {
+                                this.props.editorHandler('image-editor');
+                            }
+                        }}>
                         <FontAwesomeIcon 
                             className='insert-icon'
                             icon={faList}/>
@@ -300,8 +308,12 @@ class ToolbarContent extends React.Component
                     </button>
                     <button 
                         id='replace-image-button' 
-                        className='insert-button'
-                        onClick={() => { }}>
+                        className={replaceImageClassName}
+                        onClick={() => { 
+                            if (this.props.currentEditor === 'image-editor') {
+                                this.props.editorHandler('update-image');
+                            }
+                        }}>
                         <FontAwesomeIcon 
                             className='insert-icon'
                             icon={faFileImage}/>